requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain-anthropic==0.1.15
python-dotenv==1.0.0
google-cloud-texttospeech==2.16.3 